+++ Source/Project64-core/N64System/MemoryHandler/PeripheralInterfaceHandler.h
@@ -89,7 +89,7 @@ private:
     CartridgeDomain2Address2Handler & m_Domain2Address2Handler;
     CMipsMemoryVM & m_MMU;
     CRegisters & m_Reg;
-    uint32_t & m_PC;
+    uint64_t & m_PC;
 
     bool m_DMAUsed;
 };
